@@ -17,18 +17,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.2",
+    "@angular/animations": "^18.0.3",
     "@angular/cdk": "^17.1.2",
-    "@angular/common": "^17.1.2",
-    "@angular/compiler": "^17.1.2",
-    "@angular/core": "^17.1.2",
-    "@angular/forms": "^17.1.2",
+    "@angular/common": "^18.0.3",
+    "@angular/compiler": "^18.0.3",
+    "@angular/core": "^18.0.3",
+    "@angular/forms": "^18.0.3",
     "@angular/material": "^17.1.2",
-    "@angular/platform-browser": "^17.1.2",
-    "@angular/platform-browser-dynamic": "^17.1.2",
-    "@angular/platform-server": "^17.1.2",
-    "@angular/router": "^17.1.2",
-    "@angular/ssr": "^17.1.2",
+    "@angular/platform-browser": "^18.0.3",
+    "@angular/platform-browser-dynamic": "^18.0.3",
+    "@angular/platform-server": "^18.0.3",
+    "@angular/router": "^18.0.3",
+    "@angular/ssr": "^18.0.4",
     "bootstrap": "^5.3.2",
     "express": "^4.18.2",
     "marked": "^9.1.6",
@@ -40,14 +40,14 @@
     "zone.js": "^0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.1.2",
+    "@angular-devkit/build-angular": "^18.0.4",
     "@angular-eslint/builder": "17.2.1",
     "@angular-eslint/eslint-plugin": "17.2.1",
     "@angular-eslint/eslint-plugin-template": "17.2.1",
     "@angular-eslint/schematics": "17.2.1",
     "@angular-eslint/template-parser": "17.2.1",
-    "@angular/cli": "^17.1.2",
-    "@angular/compiler-cli": "^17.1.2",
+    "@angular/cli": "^18.0.4",
+    "@angular/compiler-cli": "^18.0.3",
     "@testing-library/angular": "^15.2.0",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jasmine-dom": "^1.3.3",
@@ -76,6 +76,6 @@
     "prettier": "^3.2.4",
     "prettier-eslint": "^16.3.0",
     "prettier-eslint-cli": "^8.0.1",
-    "typescript": "5.3.3"
+    "typescript": "5.4.5"
   }
 }
